[bd-wocfs] fix(ar,ap): tenant isolation — scope all SQL queries by tenant_id
AR v6.3.1: COALESCE(ar_customer_id, '') → ar_customer_id::TEXT in 4
lifecycle transitions, COALESCE(SUM)::BIGINT in refunds, i32 → i64 in
progress_billing test, party_client made Optional for graceful degradation.

AP v3.3.1: add tenant_id to all tax snapshot repo functions and thread
through service layer and bill approval. Two new tenant isolation
integration tests prove cross-tenant reads and writes are blocked.

+++ modules/ar/src/domain/refunds.rs
@@ -64,7 +64,7 @@ pub async fn sum_refunded<'e>(
 ) -> Result<Option<i64>, sqlx::Error> {
     sqlx::query_scalar(
         r#"
-        SELECT COALESCE(SUM(amount_cents), 0)
+        SELECT COALESCE(SUM(amount_cents)::BIGINT, 0)
         FROM ar_refunds
         WHERE charge_id = $1 AND app_id = $2 AND status IN ('pending', 'succeeded')
         "#,
